Publish 1.9.1
Fix the issue where numbers were not appearing in the detailed summary
because the view didn't have the numbers at the time it was loaded -
and that, in turn, was because a different tab of the view was
selected. The fix is to call it multiple times, in particular when the
tab changes.